@@ -1,20 +1,11 @@
 package com.group.libraryapp.dto.user.request;
 
+import lombok.Getter;
+
+@Getter
 public class UserCreateRequest {
 
     private String name;
     private Integer age; // int: null 불가 , Integer: null 가능
 
-    public UserCreateRequest(String name, Integer age) {
-        this.name = name;
-        this.age = age;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public Integer getAge() {
-        return age;
-    }
 }
